orders/src/routes/show.ts
import express, { Request, Response } from 'express';
import {
  NotAuthorizedError,
  NotFoundError,
  requireAuth
} from '@manar-mansour-org/common';
import { Order } from '../models/order';

const router = express.Router();
//we could also validate that the orderId provided is a valid mongodb id
router.get(
  '/api/orders/:orderId',
  requireAuth,
  async (req: Request, res: Response) => {
    const order = await Order.findById(req.params.orderId).populate('ticket'); //fetching the order and its associated ticket
    if (!order) {
      throw new NotFoundError();
    }
    if (order.userId !== req.currentUser!.id) {
      throw new NotAuthorizedError();
    }
    res.send(order);
  }
);

export { router as showOrderRouter };
